pyproject.toml
# Retina Image Analysis
# Python dependencies for the project

[tool.poetry]
name = "retina-image-analysis"
version = "0.1.0"
description = "AI/ML project for retinal image analysis"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
torch = {version = "^2.4.0", optional = true}
torchvision = {version = "^0.19.0", optional = true}
tensorflow = {version = "^2.14.0", optional = true}
opencv-python = "^4.8.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
pillow = "^10.0.0"
tqdm = "^4.65.0"
pyyaml = "^6.0"
mlflow = "^2.8.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pydantic = "^2.0.0"
albumentations = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"
pre-commit = "^3.3.0"
jupyter = "^1.0.0"
notebook = "^7.0.0"

[tool.poetry.group.ml.dependencies]
dvc = "^3.0.0"
wandb = "^0.16.0"
optuna = "^3.4.0"
tensorboard = "^2.14.0"

[tool.poetry.extras]
tensorflow = ["tensorflow"]
torch = ["torch", "torchvision"]
all = ["tensorflow", "torch", "torchvision"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --cov=src --cov-report=term-missing"
testpaths = [
    "tests",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
